Move pandas-specific concat logic out of the dataframe boundary module
